pyproject.toml
[tool.poetry]
name = "inv-aki-flask"
version = "0.1.0"
description = "Inverse Akinator Game for Web"
authors = ["komot <komotaymkty@gmail.com>"]
readme = "README.md"
packages = [{include = "inv_aki_flask"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.2"
gunicorn = "^20.1.0"
openai = "^0.27.8"
google-cloud-secret-manager = "^2.16.2"
google-cloud-datastore = "^2.16.1"
pandas = "^2.0.3"

[tool.poetry.scripts]
inv_aki_flask = 'inv_aki_flask.app:launch_for_local'

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B311"]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
warn_return_any = true
disallow_incomplete_defs = true
strict_optional = true
# disallow_untyped_defs = true
# disallow_any_expr = true
# disallow_any_generics = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
